fixed onsert movie functionality and added duplicate checker to prevent duplicate movies to be inserted twice based on their name. Changed column data type for "votingAverage"

+++ Database/flyway/V1.1__create_initial_tables.sql
@@ -7,7 +7,7 @@ CREATE TABLE "Movie" (
 	"overview" TEXT,
 	"releaseDate" DATE NOT NULL,
 	"poster" VARCHAR NULL,
-	"votingAverage" INT,
+	"votingAverage" DECIMAL,
 	"adult" BOOL NOT NULL DEFAULT FALSE,
 	"language" VARCHAR NOT NULL,
 	"createdTimestamp" TIMESTAMP NOT NULL DEFAULT NOW(),
